backend/requirements.txt
Django
djangorestframework
django-guardian
Pillow